requirements.txt

opencv-python
numpy
pillow
pandas
numpy
